@@ -6,8 +6,12 @@ package com.pddstudio.highlightjs.models;
  * have a look at the README.md
  */
 
+@SuppressWarnings("SpellCheckingInspection")
 public enum Theme {
+    A11Y_DARK("a11y-dark"),
+    A11Y_LIGHT("a11y-light"),
     AGATE("agate"),
+    AND_OLD_HOPE("an-old-hope"),
     ANDROID_STUDIO("androidstudio"),
     ARDUINO_LIGHT("arduino-light"),
     ARTA("arta"),
@@ -33,20 +37,23 @@ public enum Theme {
     ATELIER_SULPHURPOOL_DARK("atelier-sulphurpool-dark"),
     ATELIER_SULPHURPOOL_LIGHT("atelier-sulphurpool-light"),
     ATOM_ONE_DARK("atom-one-dark"),
+    ATOM_ONE_DARK_REASONABLE("atom-one-dark-reasonable"),
     ATOM_ONE_LIGHT("atom-one-light"),
     BROWN_PAPER("brown-paper"),
     CODEPEN_EMBED("codepen-embed"),
     COLOR_BREWER("color-brewer"),
     DARK("dark"),
-    DARKULA("darkula"),
+    DARCULA("darcula"),
     DEFAULT("default"),
     DOCCO("docco"),
-    DRAKULA("drakula"),
+    DRACULA("dracula"),
     FAR("far"),
     FOUNDATION("foundaiton"),
     GITHUB("github"),
     GITHUB_GIST("github-gist"),
+    GML("gml"),
     GOOGLECODE("googlecode"),
+    GRADIENT_DARK("gradient-dark"),
     GRAYSCALE("grayscale"),
     GRUVBOX_DARK("gruvbox-dark"),
     GRUVBOX_LIGHT("gruvbox-light"),
@@ -54,12 +61,20 @@ public enum Theme {
     HYBRID("hybrid"),
     IDEA("idea"),
     IR_BLACK("ir-black"),
+    ISBL_EDITOR_DARK("isbl-editor-dark"),
+    ISBL_EDITOR_LIGHT("isbl-editor-light"),
     KIMBIE_DARK("kimbie.dark"),
     KIMBIE_LIGHT("kimbie.light"),
+    LIGHTFAIR("lightfair"),
+    LIOSHI("lioshi"),
     MAGULA("magula"),
     MONO_BLUE("mono-blue"),
     MONOKAI("monokai"),
     MONOKAI_SUBLIME("monokai-sublime"),
+    NIGHT_OWL("night-owl"),
+    NNFX("nnfx"),
+    NNFX_DARK("nnfx-dark"),
+    NORD("nord"),
     OBSIDIAN("obsidian"),
     OCEAN("ocean"),
     PARAISO_DARK("paraiso-dark"),
@@ -69,8 +84,10 @@ public enum Theme {
     QT_CREATOR_DARK("qtcreator_dark"),
     QT_CREATOR_LIGHT("qtcreator_light"),
     RAILSCASTS("railscasts"),
-    RAINBOX("rainbow"),
+    RAINBOW("rainbow"),
+    ROUTEROS("routeros"),
     SCHOOL_BOOK("school-book"),
+    SHADES_OF_PURPLE("shades-of-purple"),
     SOLARIZED_DARK("solarized-dark"),
     SOLARIZED_LIGHT("solarized-light"),
     SUNBURST("sunburst"),
@@ -80,6 +97,7 @@ public enum Theme {
     TOMORROW_NIGHT("tomorrow-night"),
     TOMORROW_NIGHT_EIGHTIES("tomorrow-night-eighties"),
     VS("vs"),
+    VS2015("vs2015"),
     X_CODE("xcode"),
     XT256("xt256"),
     ZENBURN("zenburn");
